pyproject.toml
[project]
name = "scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dataparser>=0.0.2",
    "dateparser>=1.3.0",
    "environs>=14.5.0",
    "pika>=1.3.2",
    "pytest-playwright>=0.7.2",
]
